Get data from IntChron

IntChron is an indexing service for chronological data from multiple sources. This function queries IntChron to retrieve data from the specified database ('host'), countries and sites.

intchron(hosts, countries = NA, sites = NA, tabulate = TRUE)

intchron_hosts()

Arguments

hosts
Databases to be retrieved. See details.

countries
Vector of countries to be retrieved. See details.

sites
Vector of sites to be retrieved. See details.

tabulate
If TRUE (the default), the data retrieved will be combined into a data frame. Set FALSE to get the raw data from IntChron.

Value

At least hosts must be specified. Use intchron_hosts() to get a list of currently available hosts, or "all" to retrieve data from any database.

A data frame, or if tabulate = FALSE, a list of IntChron responses.

Details

See vignette("intchron").

Examples

# List available databases
intchron_hosts()

#>       host                             database
#> 1  egyptdb        Egyptian Radiocarbon Database
#> 2 intimate                    INTIMATE Database
#> 3     nrcf   NERC Radiocarbon Facility (Oxford)
#> 4      oxa  Oxford Radiocarbon Accelerator Unit
#> 5     sadb Southern Africa Radiocarbon Database

# Get all data from Jordan
intchron("all", countries = "Jordan")

#>                record_site record_country            record_name
#> 1             Araq ed-Dubb         Jordan           Araq ed-Dubb
#> 2              Ayn Qasiyah         Jordan            Ayn Qasiyah
#> 3              Ayn Qasiyah         Jordan            Ayn Qasiyah
#> 4              Ayn Qasiyah         Jordan            Ayn Qasiyah
#> 5              Ayn Qasiyah         Jordan            Ayn Qasiyah
#> 6              Ayn Qasiyah         Jordan            Ayn Qasiyah
#> 7              Ayn Qasiyah         Jordan            Ayn Qasiyah
#> 8              Ayn Qasiyah         Jordan            Ayn Qasiyah
#> 9              Ayn Qasiyah         Jordan            Ayn Qasiyah
#> 10             Ayn Qasiyah         Jordan            Ayn Qasiyah
#> 11             Ayn Qasiyah         Jordan            Ayn Qasiyah
#> 12                Azraq 31         Jordan               Azraq 31
#> 13                Azraq 31         Jordan               Azraq 31
#> 14                Azraq 31         Jordan               Azraq 31
#> 15               Burqu' 02         Jordan              Burqu' 02
#> 16               Burqu' 03         Jordan              Burqu' 03
#> 17               Burqu' 27         Jordan              Burqu' 27
#> 18               Burqu' 27         Jordan              Burqu' 27
#> 19               Burqu' 27         Jordan              Burqu' 27
#> 20               Burqu' 35         Jordan              Burqu' 35
#> 21               Burqu' 35         Jordan              Burqu' 35
#> 22               Burqu' 35         Jordan              Burqu' 35
#> 23  Dahikiya, Badia Region         Jordan Dahikiya, Badia Region
#> 24  Dahikiya, Badia Region         Jordan Dahikiya, Badia Region
#> 25  Dahikiya, Badia Region         Jordan Dahikiya, Badia Region
#> 26  Dahikiya, Badia Region         Jordan Dahikiya, Badia Region
#> 27                Dhuweila         Jordan               Dhuweila
#> 28                Dhuweila         Jordan               Dhuweila
#> 29                Dhuweila         Jordan               Dhuweila
#> 30                Dhuweila         Jordan               Dhuweila
#> 31             Kharaneh IV         Jordan            Kharaneh IV
#> 32             Kharaneh IV         Jordan            Kharaneh IV
#> 33             Kharaneh IV         Jordan            Kharaneh IV
#> 34             Kharaneh IV         Jordan            Kharaneh IV
#> 35             Kharaneh IV         Jordan            Kharaneh IV
#> 36             Kharaneh IV         Jordan            Kharaneh IV
#> 37             Kharaneh IV         Jordan            Kharaneh IV
#> 38             Kharaneh IV         Jordan            Kharaneh IV
#> 39             Kharaneh IV         Jordan            Kharaneh IV
#> 40             Kharaneh IV         Jordan            Kharaneh IV
#> 41             Kharaneh IV         Jordan            Kharaneh IV
#> 42             Kharaneh IV         Jordan            Kharaneh IV
#> 43             Kharaneh IV         Jordan            Kharaneh IV
#> 44             Kharaneh IV         Jordan            Kharaneh IV
#> 45           Shuna Project         Jordan          Shuna Project
#> 46           Shuna Project         Jordan          Shuna Project
#> 47           Shuna Project         Jordan          Shuna Project
#> 48           Shuna Project         Jordan          Shuna Project
#> 49           Shuna Project         Jordan          Shuna Project
#> 50           Shuna Project         Jordan          Shuna Project
#> 51           Shuna Project         Jordan          Shuna Project
#> 52           Shuna Project         Jordan          Shuna Project
#> 53           Shuna Project         Jordan          Shuna Project
#> 54           Shuna Project         Jordan          Shuna Project
#> 55           Shuna Project         Jordan          Shuna Project
#> 56           Shuna Project         Jordan          Shuna Project
#> 57           Shuna Project         Jordan          Shuna Project
#> 58           Shuna Project         Jordan          Shuna Project
#> 59           Shuna Project         Jordan          Shuna Project
#> 60           Shuna Project         Jordan          Shuna Project
#> 61           Shuna Project         Jordan          Shuna Project
#> 62           Shuna Project         Jordan          Shuna Project
#> 63           Shuna Project         Jordan          Shuna Project
#> 64           Shuna Project         Jordan          Shuna Project
#> 65           Shuna Project         Jordan          Shuna Project
#> 66           Shuna Project         Jordan          Shuna Project
#> 67           Shuna Project         Jordan          Shuna Project
#> 68           Shuna Project         Jordan          Shuna Project
#> 69           Shuna Project         Jordan          Shuna Project
#> 70           Shuna Project         Jordan          Shuna Project
#> 71           Shuna Project         Jordan          Shuna Project
#> 72           Shuna Project         Jordan          Shuna Project
#> 73           Shuna Project         Jordan          Shuna Project
#> 74           Shuna Project         Jordan          Shuna Project
#> 75           Shuna Project         Jordan          Shuna Project
#> 76           Shuna Project         Jordan          Shuna Project
#> 77           Shuna Project         Jordan          Shuna Project
#> 78           Shuna Project         Jordan          Shuna Project
#> 79           Shuna Project         Jordan          Shuna Project
#> 80           Shuna Project         Jordan          Shuna Project
#> 81           Shuna Project         Jordan          Shuna Project
#> 82           Shuna Project         Jordan          Shuna Project
#> 83           Shuna Project         Jordan          Shuna Project
#> 84           Shuna Project         Jordan          Shuna Project
#> 85           Shuna Project         Jordan          Shuna Project
#> 86           Shuna Project         Jordan          Shuna Project
#> 87           Shuna Project         Jordan          Shuna Project
#> 88           Shuna Project         Jordan          Shuna Project
#> 89      Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 90      Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 91      Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 92      Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 93      Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 94      Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 95      Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 96      Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 97      Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 98      Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 99      Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 100     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 101     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 102     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 103     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 104     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 105     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 106     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 107     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 108     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 109     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 110     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 111     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 112     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 113     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 114     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 115     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 116     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 117     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 118     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 119     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 120     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 121     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 122     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 123     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 124     Tell Abu Al-Kharaz         Jordan     Tell Abu Al-Kharaz
#> 125       Tell Abu en-Niaj         Jordan       Tell Abu en-Niaj
#> 126       Tell Abu en-Niaj         Jordan       Tell Abu en-Niaj
#> 127       Tell Abu en-Niaj         Jordan       Tell Abu en-Niaj
#> 128       Tell Abu en-Niaj         Jordan       Tell Abu en-Niaj
#> 129       Tell Abu en-Niaj         Jordan       Tell Abu en-Niaj
#> 130       Tell Abu en-Niaj         Jordan       Tell Abu en-Niaj
#> 131         Tell el-Hayyat         Jordan         Tell el-Hayyat
#> 132         Tell el-Hayyat         Jordan         Tell el-Hayyat
#> 133         Tell el-Hayyat         Jordan         Tell el-Hayyat
#> 134         Tell el-Hayyat         Jordan         Tell el-Hayyat
#> 135         Tell el-Hayyat         Jordan         Tell el-Hayyat
#> 136         Tell el-Hayyat         Jordan         Tell el-Hayyat
#> 137         Tell el-Hayyat         Jordan         Tell el-Hayyat
#> 138         Tell el-Hayyat         Jordan         Tell el-Hayyat
#> 139           Tell el-Hibr         Jordan           Tell el-Hibr
#> 140            Tell Hesban         Jordan            Tell Hesban
#> 141            Tell Hesban         Jordan            Tell Hesban
#> 142             Wadi Jilat         Jordan             Wadi Jilat
#> 143             Wadi Jilat         Jordan             Wadi Jilat
#> 144             Wadi Jilat         Jordan             Wadi Jilat
#> 145             Wadi Jilat         Jordan             Wadi Jilat
#> 146             Wadi Jilat         Jordan             Wadi Jilat
#> 147             Wadi Jilat         Jordan             Wadi Jilat
#> 148             Wadi Jilat         Jordan             Wadi Jilat
#> 149             Wadi Jilat         Jordan             Wadi Jilat
#> 150             Wadi Jilat         Jordan             Wadi Jilat
#> 151             Wadi Jilat         Jordan             Wadi Jilat
#> 152             Wadi Jilat         Jordan             Wadi Jilat
#> 153          Wadi Jilat 13         Jordan          Wadi Jilat 13
#> 154          Wadi Jilat 22         Jordan          Wadi Jilat 22
#> 155          Wadi Jilat 22         Jordan          Wadi Jilat 22
#> 156          Wadi Jilat 25         Jordan          Wadi Jilat 25
#>     record_longitude record_latitude series_type   labcode longitude latitude
#> 1            32.3167         35.6667  R_Datelist  OxA-2567   32.3167  35.6667
#> 2            36.8181         31.8339  R_Datelist OxA-18829   36.8181  31.8339
#> 3            36.8181         31.8339  R_Datelist OxA-18830   36.8181  31.8339
#> 4            36.8181         31.8339  R_Datelist OxA-18831   36.8181  31.8339
#> 5            36.8181         31.8339  R_Datelist OxA-18832   36.8181  31.8339
#> 6            36.8181         31.8339  R_Datelist OxA-18833   36.8181  31.8339
#> 7            36.8181         31.8339  R_Datelist OxA-18829   36.8181  31.8339
#> 8            36.8181         31.8339  R_Datelist OxA-18830   36.8181  31.8339
#> 9            36.8181         31.8339  R_Datelist OxA-18831   36.8181  31.8339
#> 10           36.8181         31.8339  R_Datelist OxA-18832   36.8181  31.8339
#> 11           36.8181         31.8339  R_Datelist OxA-18833   36.8181  31.8339
#> 12           36.8167         31.8000  R_Datelist   OxA-870    0.0000   0.0000
#> 13           36.8167         31.8000  R_Datelist   OxA-871    0.0000   0.0000
#> 14           36.8167         31.8000  R_Datelist  OxA-2412   36.8167  31.8000
#> 15           37.8333         32.6667  R_Datelist  OxA-2807   37.8333  32.6667
#> 16           37.8333         32.6667  R_Datelist  OxA-2808   37.8333  32.6667
#> 17           37.8333         32.6667  R_Datelist  OxA-2764   37.8333  32.6667
#> 18           37.8333         32.6667  R_Datelist  OxA-2765   37.8333  32.6667
#> 19           37.8333         32.6667  R_Datelist  OxA-2766   37.8333  32.6667
#> 20           37.8333         32.6667  R_Datelist  OxA-2768   37.8333  32.6667
#> 21           37.8333         32.6667  R_Datelist  OxA-2769   37.8333  32.6667
#> 22           37.8333         32.6667  R_Datelist  OxA-2770   37.8333  32.6667
#> 23           31.6000         37.1000  R_Datelist OxA-11594   31.6000  37.1000
#> 24           31.6000         37.1000  R_Datelist OxA-11595   31.6000  37.1000
#> 25           31.6000         37.1000  R_Datelist OxA-11594   31.6000  37.1000
#> 26           31.6000         37.1000  R_Datelist OxA-11595   31.6000  37.1000
#> 27           37.2500         32.0833  R_Datelist  OxA-1636    0.0000   0.0000
#> 28           37.2500         32.0833  R_Datelist  OxA-1637    0.0000   0.0000
#> 29           37.2500         32.0833  R_Datelist  OxA-1728   37.2500  32.0833
#> 30           37.2500         32.0833  R_Datelist  OxA-1729   37.2500  32.0833
#> 31           36.4542         31.7237  R_Datelist OxA-22273   36.4542  31.7237
#> 32           36.4542         31.7237  R_Datelist OxA-22274   36.4542  31.7237
#> 33           36.4542         31.7237  R_Datelist OxA-22275   36.4542  31.7237
#> 34           36.4542         31.7237  R_Datelist OxA-22287   36.4542  31.7237
#> 35           36.4542         31.7237  R_Datelist OxA-22288   36.4542  31.7237
#> 36           36.4542         31.7237  R_Datelist OxA-22289   36.4542  31.7237
#> 37           36.4542         31.7237  R_Datelist OxA-22290   36.4542  31.7237
#> 38           36.4542         31.7237  R_Datelist OxA-22273   36.4542  31.7237
#> 39           36.4542         31.7237  R_Datelist OxA-22274   36.4542  31.7237
#> 40           36.4542         31.7237  R_Datelist OxA-22275   36.4542  31.7237
#> 41           36.4542         31.7237  R_Datelist OxA-22287   36.4542  31.7237
#> 42           36.4542         31.7237  R_Datelist OxA-22288   36.4542  31.7237
#> 43           36.4542         31.7237  R_Datelist OxA-22289   36.4542  31.7237
#> 44           36.4542         31.7237  R_Datelist OxA-22290   36.4542  31.7237
#> 45           32.6167         35.6000  R_Datelist  OxA-4633   32.6167  35.6000
#> 46           32.6167         35.6000  R_Datelist  OxA-4634   32.6167  35.6000
#> 47           32.6167         35.6000  R_Datelist  OxA-4635   32.6167  35.6000
#> 48           32.6167         35.6000  R_Datelist  OxA-4636   32.6167  35.6000
#> 49           32.6167         35.6000  R_Datelist  OxA-4637   32.6167  35.6000
#> 50           32.6167         35.6000  R_Datelist  OxA-4638   32.6167  35.6000
#> 51           32.6167         35.6000  R_Datelist  OxA-4639   32.6167  35.6000
#> 52           32.6167         35.6000  R_Datelist  OxA-4640   32.6167  35.6000
#> 53           32.6167         35.6000  R_Datelist  OxA-4641   32.6167  35.6000
#> 54           32.6167         35.6000  R_Datelist  OxA-4642   32.6167  35.6000
#> 55           32.6167         35.6000  R_Datelist  OxA-4643   32.6167  35.6000
#> 56           32.6167         35.6000  R_Datelist  OxA-4644   32.6167  35.6000
#> 57           32.6167         35.6000  R_Datelist  OxA-5387   32.6167  35.6000
#> 58           32.6167         35.6000  R_Datelist  OxA-5388   32.6167  35.6000
#> 59           32.6167         35.6000  R_Datelist  OxA-5389   32.6167  35.6000
#> 60           32.6167         35.6000  R_Datelist  OxA-5390   32.6167  35.6000
#> 61           32.6167         35.6000  R_Datelist  OxA-5391   32.6167  35.6000
#> 62           32.6167         35.6000  R_Datelist  OxA-5392   32.6167  35.6000
#> 63           32.6167         35.6000  R_Datelist  OxA-5393   32.6167  35.6000
#> 64           32.6167         35.6000  R_Datelist  OxA-5394   32.6167  35.6000
#> 65           32.6167         35.6000  R_Datelist  OxA-5395   32.6167  35.6000
#> 66           32.6167         35.6000  R_Datelist  OxA-5396   32.6167  35.6000
#> 67           32.6167         35.6000  R_Datelist  OxA-4633   32.6167  35.6000
#> 68           32.6167         35.6000  R_Datelist  OxA-4634   32.6167  35.6000
#> 69           32.6167         35.6000  R_Datelist  OxA-4635   32.6167  35.6000
#> 70           32.6167         35.6000  R_Datelist  OxA-4636   32.6167  35.6000
#> 71           32.6167         35.6000  R_Datelist  OxA-4637   32.6167  35.6000
#> 72           32.6167         35.6000  R_Datelist  OxA-4638   32.6167  35.6000
#> 73           32.6167         35.6000  R_Datelist  OxA-4639   32.6167  35.6000
#> 74           32.6167         35.6000  R_Datelist  OxA-4640   32.6167  35.6000
#> 75           32.6167         35.6000  R_Datelist  OxA-4641   32.6167  35.6000
#> 76           32.6167         35.6000  R_Datelist  OxA-4642   32.6167  35.6000
#> 77           32.6167         35.6000  R_Datelist  OxA-4643   32.6167  35.6000
#> 78           32.6167         35.6000  R_Datelist  OxA-4644   32.6167  35.6000
#> 79           32.6167         35.6000  R_Datelist  OxA-5387   32.6167  35.6000
#> 80           32.6167         35.6000  R_Datelist  OxA-5388   32.6167  35.6000
#> 81           32.6167         35.6000  R_Datelist  OxA-5389   32.6167  35.6000
#> 82           32.6167         35.6000  R_Datelist  OxA-5390   32.6167  35.6000
#> 83           32.6167         35.6000  R_Datelist  OxA-5391   32.6167  35.6000
#> 84           32.6167         35.6000  R_Datelist  OxA-5392   32.6167  35.6000
#> 85           32.6167         35.6000  R_Datelist  OxA-5393   32.6167  35.6000
#> 86           32.6167         35.6000  R_Datelist  OxA-5394   32.6167  35.6000
#> 87           32.6167         35.6000  R_Datelist  OxA-5395   32.6167  35.6000
#> 88           32.6167         35.6000  R_Datelist  OxA-5396   32.6167  35.6000
#> 89           35.5833         32.3667  R_Datelist  OxA-4329   35.5833  32.3667
#> 90           35.5833         32.3667  R_Datelist  OxA-4330   35.5833  32.3667
#> 91           35.5833         32.3667  R_Datelist  OxA-4331   35.5833  32.3667
#> 92           35.5833         32.3667  R_Datelist  OxA-4332   35.5833  32.3667
#> 93           35.5833         32.3667  R_Datelist  OxA-4333   35.5833  32.3667
#> 94           35.5833         32.3667  R_Datelist  OxA-4334   35.5833  32.3667
#> 95           35.5833         32.3667  R_Datelist  OxA-4335   35.5833  32.3667
#> 96           35.5833         32.3667  R_Datelist  OxA-4336   35.5833  32.3667
#> 97           35.5833         32.3667  R_Datelist  OxA-4337   35.5833  32.3667
#> 98           35.5833         32.3667  R_Datelist  OxA-5088   35.5833  32.3667
#> 99           35.5833         32.3667  R_Datelist  OxA-5089   35.5833  32.3667
#> 100          35.5833         32.3667  R_Datelist  OxA-5090   35.5833  32.3667
#> 101          35.5833         32.3667  R_Datelist  OxA-5091   35.5833  32.3667
#> 102          35.5833         32.3667  R_Datelist  OxA-5092   35.5833  32.3667
#> 103          35.5833         32.3667  R_Datelist  OxA-5093   35.5833  32.3667
#> 104          35.5833         32.3667  R_Datelist  OxA-5094   35.5833  32.3667
#> 105          35.5833         32.3667  R_Datelist  OxA-5095   35.5833  32.3667
#> 106          35.5833         32.3667  R_Datelist  OxA-5096   35.5833  32.3667
#> 107          35.5833         32.3667  R_Datelist  OxA-4329   35.5833  32.3667
#> 108          35.5833         32.3667  R_Datelist  OxA-4330   35.5833  32.3667
#> 109          35.5833         32.3667  R_Datelist  OxA-4331   35.5833  32.3667
#> 110          35.5833         32.3667  R_Datelist  OxA-4332   35.5833  32.3667
#> 111          35.5833         32.3667  R_Datelist  OxA-4333   35.5833  32.3667
#> 112          35.5833         32.3667  R_Datelist  OxA-4334   35.5833  32.3667
#> 113          35.5833         32.3667  R_Datelist  OxA-4335   35.5833  32.3667
#> 114          35.5833         32.3667  R_Datelist  OxA-4336   35.5833  32.3667
#> 115          35.5833         32.3667  R_Datelist  OxA-4337   35.5833  32.3667
#> 116          35.5833         32.3667  R_Datelist  OxA-5088   35.5833  32.3667
#> 117          35.5833         32.3667  R_Datelist  OxA-5089   35.5833  32.3667
#> 118          35.5833         32.3667  R_Datelist  OxA-5090   35.5833  32.3667
#> 119          35.5833         32.3667  R_Datelist  OxA-5091   35.5833  32.3667
#> 120          35.5833         32.3667  R_Datelist  OxA-5092   35.5833  32.3667
#> 121          35.5833         32.3667  R_Datelist  OxA-5093   35.5833  32.3667
#> 122          35.5833         32.3667  R_Datelist  OxA-5094   35.5833  32.3667
#> 123          35.5833         32.3667  R_Datelist  OxA-5095   35.5833  32.3667
#> 124          35.5833         32.3667  R_Datelist  OxA-5096   35.5833  32.3667
#> 125          35.5833         32.4167  R_Datelist OxA-10990   35.5833  32.4167
#> 126          35.5833         32.4167  R_Datelist OxA-10991   35.5833  32.4167
#> 127          35.5833         32.4167  R_Datelist OxA-10992   35.5833  32.4167
#> 128          35.5833         32.4167  R_Datelist OxA-10990   35.5833  32.4167
#> 129          35.5833         32.4167  R_Datelist OxA-10991   35.5833  32.4167
#> 130          35.5833         32.4167  R_Datelist OxA-10992   35.5833  32.4167
#> 131          35.5833         32.4167  R_Datelist OxA-10986   35.5833  32.4167
#> 132          35.5833         32.4167  R_Datelist OxA-10987   35.5833  32.4167
#> 133          35.5833         32.4167  R_Datelist OxA-10988   35.5833  32.4167
#> 134          35.5833         32.4167  R_Datelist OxA-10989   35.5833  32.4167
#> 135          35.5833         32.4167  R_Datelist OxA-10986   35.5833  32.4167
#> 136          35.5833         32.4167  R_Datelist OxA-10987   35.5833  32.4167
#> 137          35.5833         32.4167  R_Datelist OxA-10988   35.5833  32.4167
#> 138          35.5833         32.4167  R_Datelist OxA-10989   35.5833  32.4167
#> 139          38.1667         32.0000  R_Datelist  OxA-2767   38.1667  32.0000
#> 140          35.1333         31.1333  R_Datelist OxA-22099   35.1333  31.1333
#> 141          35.1333         31.1333  R_Datelist OxA-22099   35.1333  31.1333
#> 142          36.4167         31.5000  R_Datelist  OxA-1770    0.0000   0.0000
#> 143          36.4167         31.5000  R_Datelist  OxA-1771    0.0000   0.0000
#> 144          36.4167         31.5000  R_Datelist  OxA-1772    0.0000   0.0000
#> 145          36.4167         31.5000  R_Datelist  OxA-1799    0.0000   0.0000
#> 146          36.4167         31.5000  R_Datelist  OxA-1800    0.0000   0.0000
#> 147          36.4167         31.5000  R_Datelist  OxA-1801    0.0000   0.0000
#> 148          36.4167         31.5000  R_Datelist  OxA-1802    0.0000   0.0000
#> 149          36.4167         31.5000  R_Datelist  OxA-1812    0.0000   0.0000
#> 150          36.4167         31.5000  R_Datelist  OxA-2407    0.0000   0.0000
#> 151          36.4167         31.5000  R_Datelist  OxA-2413    0.0000   0.0000
#> 152          36.4167         31.5000  R_Datelist  OxA-2969   36.4167  31.5000
#> 153          36.4167         31.5000  R_Datelist  OxA-2411   36.4167  31.5000
#> 154          36.4167         31.5000  R_Datelist  OxA-2409   36.4167  31.5000
#> 155          36.4167         31.5000  R_Datelist  OxA-2410   36.4167  31.5000
#> 156          36.4167         31.5000  R_Datelist  OxA-2408   36.4167  31.5000
#>                 sample      material                   species    d13C r_date
#> 1                WY130      charcoal                         - -26.100   9950
#> 2      Sample Number 8      charcoal Amygdalus (possibly almon -25.859  17550
#> 3      Sample Number 8      charcoal Amygdalus (possibly almon -26.004  17490
#> 4           Sample #33      charcoal            chenopodiaceae -11.272  17555
#> 5           Sample #24      charcoal                 Amygdalus -23.747  17495
#> 6         Sample #2009      charcoal           Chenopodiaeceae -11.412   9195
#> 7      Sample Number 8      charcoal Amygdalus (possibly almon -25.859  17550
#> 8      Sample Number 8      charcoal Amygdalus (possibly almon -26.004  17490
#> 9           Sample #33      charcoal            chenopodiaceae -11.272  17555
#> 10          Sample #24      charcoal                 Amygdalus -23.747  17495
#> 11        Sample #2009      charcoal           Chenopodiaeceae -11.412   9195
#> 12            Sq.1, 10      charcoal                         -      NA   8350
#> 13          Sq.14, 107          bone                         -      NA   1280
#> 14               C,19b      charcoal                         - -10.500   8275
#> 15             B.02203      charcoal                         - -13.700      0
#> 16             B.03158      charcoal                         - -11.900   6900
#> 17             B.27132      charcoal                         - -13.800   7270
#> 18             B.27141      charcoal                         - -13.600   7350
#> 19             B.27142      charcoal                         - -12.500   7930
#> 20             B.35207      charcoal                         - -10.800   8140
#> 21             B.35208      charcoal                         - -12.200   8180
#> 22             B.35112      charcoal                         - -11.200   8270
#> 23             BRT-DA1 plant remains                           -11.487      0
#> 24             BRT-DA2 plant remains                           -11.571      0
#> 25             BRT-DA1 plant remains                           -11.487      0
#> 26             BRT-DA2 plant remains                           -11.571      0
#> 27           2202/4217      charcoal                         - -26.000   7030
#> 28           2202/2133      charcoal                         - -26.000   8350
#> 29                4225      charcoal                         - -26.000   7140
#> 30                4118      charcoal                         - -26.000   7450
#> 31        KHIV-AreaB-6      charcoal            Chenopodiaceae -23.846  15890
#> 32       KHIV-AreaB-10      charcoal             Indet. Dicot. -11.040  15770
#> 33       KHIV-AreaB-19      charcoal             Indet. Dicot. -26.070  16145
#> 34        KHIV-AreaB-4      charcoal            Chenopodiaceae -22.959  15980
#> 35       KHIV-AreaB-15      charcoal            Chenopodiaceae -24.045  16275
#> 36       KHIV-AreaB-24      charcoal             Indet. Dicot. -10.156  16300
#> 37       KHIV-AreaB-25      charcoal            Chenopodiaceae -24.492  16200
#> 38        KHIV-AreaB-6      charcoal            Chenopodiaceae -23.846  15890
#> 39       KHIV-AreaB-10      charcoal             Indet. Dicot. -11.040  15770
#> 40       KHIV-AreaB-19      charcoal             Indet. Dicot. -26.070  16145
#> 41        KHIV-AreaB-4      charcoal            Chenopodiaceae -22.959  15980
#> 42       KHIV-AreaB-15      charcoal            Chenopodiaceae -24.045  16275
#> 43       KHIV-AreaB-24      charcoal             Indet. Dicot. -10.156  16300
#> 44       KHIV-AreaB-25      charcoal            Chenopodiaceae -24.492  16200
#> 45                  11      charcoal                   Tamarix -25.990   4500
#> 46                  12      charcoal                  Ulmaceae -26.980   4440
#> 47                  13      charcoal                   Tamarix -27.250   4520
#> 48                  14      charcoal                       Ash -23.500   4555
#> 49                  15      charcoal                     Olive -22.770   4920
#> 50                  16      charcoal                     Olive -22.310   4870
#> 51                  17      charcoal                     Olive -24.350   4975
#> 52                  18      charcoal                   Quercus -24.940   4820
#> 53                  11 charred seeds                     Olive -23.120   5020
#> 54                  11 charred seeds                     Olive -22.280   5080
#> 55                  12 charred seeds             hulled barley -23.820   6055
#> 56                  12 charred seeds             hulled barley -23.660   6000
#> 57      9, Area D: 443      charcoal                   tamarix -25.400   6020
#> 58      10, Area D:585         seeds           Hordeum sativum -22.100   6040
#> 59          13, Ac 725      charcoal                  tamarisk -26.310   4750
#> 60          14, Aa:813         seeds                    lentil -22.460   4665
#> 61          15, Aa:626         seeds               olive stone -21.320   4585
#> 62          16, Aa:696         seeds                     emmer -20.950   4590
#> 63          17, Ab:744         seeds             hulled barley -19.770   4685
#> 64          18, Ab:753         seeds                     emmer -23.170   4735
#> 65          19, Ab:869         seeds               olive stone -22.880   6065
#> 66          20, Ab:841         seeds                     emmer -22.900   6070
#> 67                  11      charcoal                   Tamarix -25.990   4500
#> 68                  12      charcoal                  Ulmaceae -26.980   4440
#> 69                  13      charcoal                   Tamarix -27.250   4520
#> 70                  14      charcoal                       Ash -23.500   4555
#> 71                  15      charcoal                     Olive -22.770   4920
#> 72                  16      charcoal                     Olive -22.310   4870
#> 73                  17      charcoal                     Olive -24.350   4975
#> 74                  18      charcoal                   Quercus -24.940   4820
#> 75                  11 charred seeds                     Olive -23.120   5020
#> 76                  11 charred seeds                     Olive -22.280   5080
#> 77                  12 charred seeds             hulled barley -23.820   6055
#> 78                  12 charred seeds             hulled barley -23.660   6000
#> 79      9, Area D: 443      charcoal                   tamarix -25.400   6020
#> 80      10, Area D:585         seeds           Hordeum sativum -22.100   6040
#> 81          13, Ac 725      charcoal                  tamarisk -26.310   4750
#> 82          14, Aa:813         seeds                    lentil -22.460   4665
#> 83          15, Aa:626         seeds               olive stone -21.320   4585
#> 84          16, Aa:696         seeds                     emmer -20.950   4590
#> 85          17, Ab:744         seeds             hulled barley -19.770   4685
#> 86          18, Ab:753         seeds                     emmer -23.170   4735
#> 87          19, Ab:869         seeds               olive stone -22.880   6065
#> 88          20, Ab:841         seeds                     emmer -22.900   6070
#> 89    1/K91 VIIC GL140      charcoal                         - -23.200   4475
#> 90      2/K92 XIB 4L55      charcoal                         - -23.700   4440
#> 91    3/K91 VIII 5L144      charcoal                         - -20.900   4445
#> 92      4/K91 IX 3L223      charcoal                         - -21.700   4600
#> 93      K92 XIB 6 L110      charcoal                         - -19.200   4360
#> 94  6/K91 VIIB 12 L230      charcoal                         - -24.200   4435
#> 95   7/K92 VIIC 8 L281      charcoal                         - -21.600   4405
#> 96  8/K91 VIIA 12 L246      charcoal                         - -25.100   4540
#> 97      9/K92 XIC 6L93  charred bone                         - -22.100   2910
#> 98    1/K93 XXII 2 L42      charcoal                           -24.570   2495
#> 99  2/K93 VIII A4 L380 charred seeds                           -23.230   3260
#> 100 3/K93 VIII A4 L368 charred seeds                           -23.350   3210
#> 101    4/K93 IX 5 L322 charred seeds                           -22.840   4390
#> 102 5/K93 VIIIB 7 L412 charred seeds                           -25.100   4265
#> 103 6/K93 VIIIA 8 L416      charcoal                           -25.100   4450
#> 104 7/K93 VIIIA 9 L441 charred seeds                           -22.410   4400
#> 105 8/K93 VIIIA 9 L444      charcoal                           -24.570   4440
#> 106 9/K93 VIIIb 9 L434      charcoal                           -21.970   4335
#> 107   1/K91 VIIC GL140      charcoal                         - -23.200   4475
#> 108     2/K92 XIB 4L55      charcoal                         - -23.700   4440
#> 109   3/K91 VIII 5L144      charcoal                         - -20.900   4445
#> 110     4/K91 IX 3L223      charcoal                         - -21.700   4600
#> 111     K92 XIB 6 L110      charcoal                         - -19.200   4360
#> 112 6/K91 VIIB 12 L230      charcoal                         - -24.200   4435
#> 113  7/K92 VIIC 8 L281      charcoal                         - -21.600   4405
#> 114 8/K91 VIIA 12 L246      charcoal                         - -25.100   4540
#> 115     9/K92 XIC 6L93  charred bone                         - -22.100   2910
#> 116   1/K93 XXII 2 L42      charcoal                           -24.570   2495
#> 117 2/K93 VIII A4 L380 charred seeds                           -23.230   3260
#> 118 3/K93 VIII A4 L368 charred seeds                           -23.350   3210
#> 119    4/K93 IX 5 L322 charred seeds                           -22.840   4390
#> 120 5/K93 VIIIB 7 L412 charred seeds                           -25.100   4265
#> 121 6/K93 VIIIA 8 L416      charcoal                           -25.100   4450
#> 122 7/K93 VIIIA 9 L441 charred seeds                           -22.410   4400
#> 123 8/K93 VIIIA 9 L444      charcoal                           -24.570   4440
#> 124 9/K93 VIIIb 9 L434      charcoal                           -21.970   4335
#> 125          C14MBA#43         seeds           Hordeum vulgare -23.335   3932
#> 126          C14MBA#44         seeds           Hordeum vulgare -22.415   3877
#> 127          C14MBA#45         seeds                    cereal -22.830   3886
#> 128          C14MBA#43         seeds           Hordeum vulgare -23.335   3932
#> 129          C14MBA#44         seeds           Hordeum vulgare -22.415   3877
#> 130          C14MBA#45         seeds                    cereal -22.830   3886
#> 131          C14MBA#39         seeds         Triticum aestivum -22.443   3470
#> 132          C14MBA#40         seeds         Triticum aestivum -22.879   3497
#> 133          C14MBA#41         seeds          olive stone/olea -21.280   3502
#> 134          C14MBA#42         seeds          olive stone/olea -21.290   3523
#> 135          C14MBA#39         seeds         Triticum aestivum -22.443   3470
#> 136          C14MBA#40         seeds         Triticum aestivum -22.879   3497
#> 137          C14MBA#41         seeds          olive stone/olea -21.280   3502
#> 138          C14MBA#42         seeds          olive stone/olea -21.290   3523
#> 139           HIBR 108      charcoal                         - -12.300   3950
#> 140       H74-D2-MM408          bone            Cervus elaphus -20.653    864
#> 141       H74-D2-MM408          bone            Cervus elaphus -20.653    864
#> 142         22 1 6b 13      charcoal                         - -26.000  11920
#> 143        22 3 15b 37      charcoal                         - -26.000  13040
#> 144        22 3 15b 36      charcoal                         - -26.000  12840
#> 145            7 A 34b      charcoal                         - -26.000   5840
#> 146        13 A 21a 19      charcoal                         - -26.000   7920
#> 147         13 A 15a 9      charcoal                         - -26.000   7870
#> 148         26 Cb 7a 1      charcoal                         - -26.000   8690
#> 149         27 E 41a 7      charcoal                           -26.000   5270
#> 150         26 C 12a 3      charcoal                         - -12.800   8720
#> 151           7, A 34a      charcoal                         - -23.400   8390
#> 152          26,ED,18a      charcoal                         - -11.200   8740
#> 153               C,24      charcoal                         - -24.200   7900
#> 154              17a,1      charcoal                         - -15.200  13490
#> 155              171,2      charcoal                         - -13.100  13540
#> 156 Area Aa locus 10aV      charcoal                         - -11.100   8020
#>     r_date_sigma qual     F14C F14C_sigma
#> 1            100      0.289778 0.00360734
#> 2             75      0.112520 0.00108000
#> 3             75      0.113330 0.00104000
#> 4             75      0.112460 0.00102000
#> 5             70      0.113260 0.00099000
#> 6             40      0.318350 0.00166000
#> 7             75      0.112520 0.00108000
#> 8             75      0.113330 0.00104000
#> 9             75      0.112460 0.00102000
#> 10            70      0.113260 0.00099000
#> 11            40      0.318350 0.00166000
#> 12           120      0.353645 0.00528288
#> 13            90      0.852704 0.00955351
#> 14            80      0.356962 0.00355496
#> 15             0    m 1.442000 0.01000000
#> 16           100      0.423604 0.00527329
#> 17            80      0.404535 0.00402873
#> 18            80      0.400526 0.00398881
#> 19            80      0.372627 0.00371096
#> 20            90      0.363012 0.00406711
#> 21            80      0.361209 0.00359725
#> 22            80      0.357184 0.00355717
#> 23             0    m 1.173600 0.00380000
#> 24             0    m 1.042300 0.00370000
#> 25             0    m 1.173600 0.00380000
#> 26             0    m 1.042300 0.00370000
#> 27            90      0.416804 0.00466978
#> 28           100      0.353645 0.00440240
#> 29            90      0.411135 0.00460627
#> 30            90      0.395571 0.00443189
#> 31            90      0.138350 0.00147000
#> 32            80      0.140330 0.00146000
#> 33            75      0.134010 0.00125000
#> 34            60      0.136790 0.00100000
#> 35            60      0.131870 0.00098000
#> 36            65      0.131430 0.00102000
#> 37            65      0.133080 0.00106000
#> 38            90      0.138350 0.00147000
#> 39            80      0.140330 0.00146000
#> 40            75      0.134010 0.00125000
#> 41            60      0.136790 0.00100000
#> 42            60      0.131870 0.00098000
#> 43            65      0.131430 0.00102000
#> 44            65      0.133080 0.00106000
#> 45           120      0.570800 0.00860000
#> 46            80      0.575200 0.00590000
#> 47            75      0.569700 0.00530000
#> 48            70      0.567200 0.00500000
#> 49            75      0.542200 0.00490000
#> 50           150      0.545100 0.01050000
#> 51            75      0.538200 0.00500000
#> 52            80      0.549000 0.00560000
#> 53            75      0.535300 0.00500000
#> 54            75      0.531200 0.00500000
#> 55            80      0.470600 0.00460000
#> 56            75      0.473900 0.00440000
#> 57            65      0.472700 0.00380000
#> 58            55      0.471400 0.00320000
#> 59            55      0.553700 0.00380000
#> 60            55      0.559600 0.00390000
#> 61            50      0.564900 0.00350000
#> 62            50      0.564600 0.00340000
#> 63            75      0.558100 0.00520000
#> 64            55      0.554800 0.00370000
#> 65            55      0.470000 0.00310000
#> 66            55      0.469700 0.00320000
#> 67           120      0.570800 0.00860000
#> 68            80      0.575200 0.00590000
#> 69            75      0.569700 0.00530000
#> 70            70      0.567200 0.00500000
#> 71            75      0.542200 0.00490000
#> 72           150      0.545100 0.01050000
#> 73            75      0.538200 0.00500000
#> 74            80      0.549000 0.00560000
#> 75            75      0.535300 0.00500000
#> 76            75      0.531200 0.00500000
#> 77            80      0.470600 0.00460000
#> 78            75      0.473900 0.00440000
#> 79            65      0.472700 0.00380000
#> 80            55      0.471400 0.00320000
#> 81            55      0.553700 0.00380000
#> 82            55      0.559600 0.00390000
#> 83            50      0.564900 0.00350000
#> 84            50      0.564600 0.00340000
#> 85            75      0.558100 0.00520000
#> 86            55      0.554800 0.00370000
#> 87            55      0.470000 0.00310000
#> 88            55      0.469700 0.00320000
#> 89            60      0.572881 0.00427896
#> 90            60      0.575383 0.00429764
#> 91            60      0.575025 0.00429497
#> 92            65      0.564036 0.00456396
#> 93            60      0.581141 0.00434066
#> 94            65      0.575741 0.00465868
#> 95            65      0.577895 0.00467611
#> 96            70      0.568264 0.00495189
#> 97            65      0.696104 0.00563261
#> 98            45      0.732900 0.00420000
#> 99            50      0.666500 0.00400000
#> 100           60      0.670700 0.00510000
#> 101           60      0.579100 0.00430000
#> 102           55      0.587900 0.00420000
#> 103           60      0.574600 0.00420000
#> 104           50      0.578300 0.00370000
#> 105           60      0.575400 0.00430000
#> 106           55      0.582900 0.00400000
#> 107           60      0.572881 0.00427896
#> 108           60      0.575383 0.00429764
#> 109           60      0.575025 0.00429497
#> 110           65      0.564036 0.00456396
#> 111           60      0.581141 0.00434066
#> 112           65      0.575741 0.00465868
#> 113           65      0.577895 0.00467611
#> 114           70      0.568264 0.00495189
#> 115           65      0.696104 0.00563261
#> 116           45      0.732900 0.00420000
#> 117           50      0.666500 0.00400000
#> 118           60      0.670700 0.00510000
#> 119           60      0.579100 0.00430000
#> 120           55      0.587900 0.00420000
#> 121           60      0.574600 0.00420000
#> 122           50      0.578300 0.00370000
#> 123           60      0.575400 0.00430000
#> 124           55      0.582900 0.00400000
#> 125           38      0.613000 0.00290000
#> 126           40      0.617100 0.00300000
#> 127           40      0.616400 0.00300000
#> 128           38      0.613000 0.00290000
#> 129           40      0.617100 0.00300000
#> 130           40      0.616400 0.00300000
#> 131           36      0.649200 0.00290000
#> 132           37      0.647000 0.00300000
#> 133           37      0.646600 0.00290000
#> 134           39      0.645000 0.00310000
#> 135           36      0.649200 0.00290000
#> 136           37      0.647000 0.00300000
#> 137           37      0.646600 0.00290000
#> 138           39      0.645000 0.00310000
#> 139           80      0.611573 0.00609060
#> 140           28      0.898030 0.00308000
#> 141           28      0.898030 0.00308000
#> 142          180      0.226756 0.00508106
#> 143          180      0.197246 0.00441980
#> 144          140      0.202218 0.00352429
#> 145          100      0.483356 0.00601713
#> 146          100      0.373091 0.00464448
#> 147          100      0.375420 0.00467348
#> 148          110      0.338989 0.00464195
#> 149           90      0.518900 0.00581365
#> 150          100      0.337725 0.00420422
#> 151           80      0.351888 0.00350443
#> 152          110      0.336886 0.00461315
#> 153           80      0.374021 0.00372485
#> 154          110      0.186500 0.00255384
#> 155          120      0.185343 0.00276872
#> 156           80      0.368475 0.00366962