Query IntChron

IntChron is an indexing service for chronological data from multiple sources. This function gets records from databases ('hosts') indexed by IntChron, optionally filtering by country and/or site.

intchron(hosts, countries = NA, sites = NA, tabulate = TRUE)

Arguments

hosts
Vector of databases to be retrieved, or "all".

countries
(Optional) Vector of countries to be retrieved.

sites
(Optional) Vector of sites to be retrieved.

tabulate
If TRUE (the default), the data retrieved will be combined into a data frame. Set FALSE to get the raw data from IntChron.

Value

A tibble, or if tabulate = FALSE, a list, of IntChron responses.

Details

At least hosts must be specified. Use intchron_hosts() to get a list of currently available hosts, or "all" to retrieve data from any database.

See vignette("rintchron") for further details.

See also

Examples

# Get data from Jordan from the ORAU and NERC databases
intchron(c("oxa", "nrcf"), countries = "Jordan")

#> # A tibble: 156 x 19
#>    record_site record_country record_name record_longitude record_latitude
#>    <chr>       <chr>          <chr>                  <dbl>           <dbl>
#>  1 Araq ed-Du… Jordan         Araq ed-Du…             32.3            35.7
#>  2 Ayn Qasiyah Jordan         Ayn Qasiyah             36.8            31.8
#>  3 Ayn Qasiyah Jordan         Ayn Qasiyah             36.8            31.8
#>  4 Ayn Qasiyah Jordan         Ayn Qasiyah             36.8            31.8
#>  5 Ayn Qasiyah Jordan         Ayn Qasiyah             36.8            31.8
#>  6 Ayn Qasiyah Jordan         Ayn Qasiyah             36.8            31.8
#>  7 Azraq 31    Jordan         Azraq 31                36.8            31.8
#>  8 Azraq 31    Jordan         Azraq 31                36.8            31.8
#>  9 Azraq 31    Jordan         Azraq 31                36.8            31.8
#> 10 Burqu' 02   Jordan         Burqu' 02               37.8            32.7
#> # … with 146 more rows, and 14 more variables: series_type <chr>,
#> #   labcode <chr>, longitude <dbl>, latitude <dbl>, sample <chr>,
#> #   material <chr>, species <chr>, d13C <dbl>, r_date <int>,
#> #   r_date_sigma <int>, qual <chr>, F14C <dbl>, F14C_sigma <dbl>, refs <chr>